requirements.txt
gradio
gradio[oauth]
requests
urllib3
python-dotenv
ipykernel

huggingface_hub
transformers

langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-groq
langchain_openai
langchain-tavily
langchain-chroma
langgraph

pandas
matplotlib

supabase
arxiv
pymupdf
wikipedia
pgvector
pillow
pytesseract
